metrics/requirements.txt
libclang>=16.0.0
regex>=2023.3.23

